+++ NEWS.md
@@ -1,6 +1,7 @@
 # recipes 0.1.1
 
 * `step_unorder` converts ordered factors to unordered factors. 
+* `step_count` counts the number of instances that a pattern exists in a string. 
 * [`dplyr::one_of`](https://github.com/topepo/recipes/issues/85) was added to the list of selectors 
 
 # recipes 0.1.0